Add files via upload
Assignment #2 - E-R Diagrams [Supertypes & Subtypes]
1.	A nonprofit organization depends on a number of different types of persons for its successful operation.  The organization is interested in the following attributes for all these persons: SSN, Name, Address, Telephone_No.  Three types of persons are of greatest interest: employees, volunteers, and donors.  Employees only have a Date_Hired attribute, and volunteers only have a Skill attribute.  Donors only have a relationship (named donates) with an Item entity type.  A donor must have donated one or more items, and an item may have no donors, or one or more donors. 
There are persons other than employees, volunteers and donors who are of interest to the organization, so that a person need not belong to any of these groups. On the other hand, a person may belong to two or more groups (e.g., employee, donor). 
2.	A rental car agency classifies the vehicles it rents into four categories: compact, mid-size, full-size, and sport utility.  The agency wants to record the following data for all vehicles: VID, Make, Model, Year, and Color.  There are no unique attributes for any of the four classes of the vehicles.  The entity type vehicle has a relationship (named rents) with a customer entity type.  None of the four vehicle classes has a unique relationship with any entity type.   

3.	A bank has three types of accounts: checking, savings, and loan.  Following are the attributes for each type of account. Assume that each bank account must be a member of exactly one of these subtypes. 
CHECKING: Acct_No, Date_Opened, Balance, Service_Charge 
SAVINGS: Acct_No, Date_Opened, Balance, Interest_Rate
LOAN: Acct_No, Date_Opened, Balance, Interest_Rate, Payment 
4.	Define an ERD for the following narrative. The database should track homes, owners, agents, offices, buyers and offers. 

a.	A home has a unique home identifier, a street address, a city, a state, a zip, a number of bedrooms, a number of bathrooms, and square feet. A home is either owner-occupied or rented. 
b.	An owner has a social security number, a name, an optional spouse name, a profession, and an optional spouse profession. An owner can possess one or more homes. Each home has only one owner. 
c.	Agents represent owners in the sale of a home. An agent can list many homes but only one agent can list a home. An agent has a unique agent identifier, a name, and a phone number. When an owner agrees to list a home with an agent, a commission (percentage of the sales price) and listing price are determined. 
d.	Each agent has an office. Data about an office include the phone number, the manager name, and the address. 
e.	A buyer has a social security number, a name, a phone, preferences for the number of bedrooms and bathrooms, and a price range. An agent can work with many buyers, but a buyer works with only one agent. Define a generalization hierarchy to depict similarities between buyers and owners. 
f.	A buyer makes an offer on a home for a specified sales price. The offer starts on the submission date and time and expires on the specified date and time. A unique offer number identifies an offer. A buyer may submit multiple offers for the same home.